+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Enhancing core SCSS and adding a new set of mixins.
 - Enhancing router by adding support for endpoint query vars.
 - Adding new Facade that makes it easier to encode/decode a single ID.
+- Adding `Paginator{}` class and utilities.
 
 ## v160831.61689